refactor(security): split lib.rs into lib, scanner, helpers, and tests modules
Same shape as the earlier refactors: the 1520-line crates/security/src/lib.rs
is split into four modules. All existing unit tests still pass.
